+++ py/desispec/qa/qa_frame.py
@@ -8,6 +8,8 @@ import numpy as np
 import os
 
 from desiutil.log import get_logger
+from desispec.io.params import read_obj_param
+
 
 # log=get_logger()
 
@@ -44,7 +46,8 @@ class QA_Frame(object):
             self.qa_data = {}
 
         # Final test
-        assert self.flavor in ['none', 'flat', 'arc', 'dark', 'bright', 'bgs', 'mws', 'lrg', 'elg', 'qso', 'gray']
+        obj_params = read_obj_param()
+        assert self.flavor in obj_params['frame_types']
 
     def init_qatype(self, qatype, param, re_init=False):
         """Initialize parameters for a given qatype
@@ -99,7 +102,9 @@ class QA_Frame(object):
 
         """
         log=get_logger()
-        assert self.flavor in ['dark','bright','bgs','mws','lrg','elg','qso','gray']
+        obj_params = read_obj_param()
+        assert self.flavor in obj_params['sci_types']
+        #assert self.flavor in ['dark','bright','bgs','mws','lrg','elg','qso','gray']
 
         # Standard FLUXCALIB input parameters
         flux_dict = dict(ZP_WAVE=0.,        # Wavelength for ZP evaluation (camera dependent)
@@ -127,8 +132,9 @@ class QA_Frame(object):
         re_init: bool, (optional)
           Re-initialize SKYSUB parameter dict
         """
-        #
-        assert self.flavor in ['dark','bright','bgs','mws','lrg','elg','qso','gray']
+        obj_params = read_obj_param()
+        assert self.flavor in obj_params['sci_types']
+        #assert self.flavor in ['dark','bright','bgs','mws','lrg','elg','qso','gray']
 
         # Standard SKYSUB input parameters
         sky_dict = dict(