+++ doc/changes.rst
@@ -13,6 +13,8 @@ desispec Change Log
 * Added function to calculate BRICKID at a given location (PR `#378`_).
 * Additional LOCATION, DEVICE_LOC, and PETAL_LOC columns for fibermap (PR `#379`_).
 * Create util.py in tests/ which is intended to contain methods to facilitate test runs
+* Add vette() method for Frame class
+* Began a desispec parameter file:  data/params/desispec_param.yaml
 
 .. _`#378`: https://github.com/desihub/desispec/pull/378
 .. _`#379`: https://github.com/desihub/desispec/pull/379